fix(restart): never report a failed restart as a success
`sac agents restart` printed green "Agent '<name>' restarted" on the line
immediately after "FAIL: duplicate session". The operator hit this live on
neurovista, believed it had relaunched on freshly-picked credentials, was in
fact still talking to the OLD process holding the OLD token, saw "Login
expired", and went diagnosing a credential store that was entirely healthy.
A tool that reports a failure as a pass sends its user to the wrong subsystem.

The lie originates at the bottom, not in the CLI:

* `tui_session`'s duplicate-session guard REFUSES to start over a live session
  and then `return True`. Idempotent-True is defensible for a plain
  `sac agents start` (an already-running agent is "fine, it's up"), but a
  RESTART's contract is to REPLACE the process, so that verdict is false.
  `agent_start` returned success, `agent_restart` propagated it, and the CLI
  faithfully printed it. Three honest layers fed one lie.

Fixes:

* `agent_restart` now passes `force=True` to its start leg, so a restart can
  actually take over a session that survived SIGTERM (the exact case: the stop
  leg warns "SIGTERM ignored ... proceeding to start anyway", the stale session
  lives, start no-ops, and the restart silently does nothing).
* the CLI stops DISCARDING `agent_restart`'s return value, and reports failure
  in red with the remedy that works. The host-listen branch likewise printed
  green "restarted" even on a non-zero returncode.
* cross-host restarts now trust the PEER'S OWN `restarted` verdict rather than
  assuming ssh-exited-0 means the restart worked -- otherwise the same false
  success simply relocates one hop away.
* the duplicate-session guard no longer suggests `sac agents restart` as the
  remedy: when it fires DURING a restart that is the very command that just
  failed. It now gives `tmux kill-session` + `start --fresh`.

Refactor (required by the 512-line limit, done properly rather than bypassed):
cross-host dispatch + host-listen bypass move to `_restart_remote.py`;
`_restart.py` re-exports them so existing imports keep resolving (542 -> 449).

Reported by scitex-writer from the operator's live session.
