+++ src/scitex_agent_container/_lifecycle/_stop.py
@@ -398,10 +398,31 @@ def agent_restart(
     # This does NOT weaken the human-at-a-TTY guard: a bare ``sac agents
     # start``/``restart`` with no consent still refuses — only the
     # pre-authorized restart's own start leg asserts the consent already given.
+    # ``force=True`` — a RESTART's contract is to REPLACE the process, so its
+    # start leg must be allowed to take over a surviving session.
+    #
+    # Without force, the start leg hits ``tui_session``'s duplicate-session
+    # guard, which is *idempotent for a plain `sac agents start`* (an
+    # already-running agent is "fine, it's up") and therefore RETURNS TRUE. For
+    # a restart that verdict is a LIE: the caller asked for a new process and
+    # got the old one. The failure mode is not theoretical — it happens exactly
+    # when the stop leg above could not kill the old runtime (SIGTERM ignored,
+    # ``_wait_for_previous_runtime_to_exit`` returned False and we proceeded
+    # anyway). Then: stale session survives -> start no-ops -> returns True ->
+    # the CLI prints "Agent '<name>' restarted". The operator hit precisely
+    # this on neurovista: he believed it had relaunched on freshly-picked
+    # credentials, was in fact still talking to the OLD process on its OLD
+    # token, saw "Login expired", and went diagnosing a credential store that
+    # was entirely healthy.
+    #
+    # So a restart FORCES: the force branch tears the stale session down first,
+    # which both makes the restart actually happen and makes a genuine failure
+    # report as a failure.
     return agent_start(
         config_path,
         registry,
         assume_yes=True,
+        force=True,
         runtime_factory=runtime_factory,
         sleep_fn=sleep_fn,
         handover_mod=handover_mod,
